* master: (54 commits)
  TextView: Fix bug in rtrim_if when entire view is cleared.
  Step 23: Renames Rollback object -> FileManager consistently
  Step 22: Renames Rollback.cc/.h -> ConfigManager.cc/.h
  Step 21: Removes the UNVERSIONED option flag completely
  Step 20: Removes incVersion/inc_version from Filemanager etc.
  Step 19: Removes the now unused Rollback member variables
  Step 18: Removes setLastModifiedTime, and simplifies code
  Step 17: Removes versionInfo, as well as currentVersion and highestSeen
  Step 16: Removes createPathStr() and cleans up statFile()
  Step 15: Removes getBaseName() and various unused members
  Step 14: Removes numberOfVersions() and getCurrentVersion()
  Step 13: Removes openFile() and closeFile()
  Step 12: Removes extractVersionInfo()
  Step 11: Removes statVersion()
  Step 10: Removes versionTimeStamp() and versionTimeStamp_ml()
  Step 9: Removes getVersion() and getVersion_ml()
  Step 8: Removes internalUpdate(), functionality was moved in 6
  Step 7: Removes updateVersion() and updateVersion_ml()
  Step 6: Simplifies checkForUserUpdate(), moving critical code here
  Step 5: Remove removeVersion() and removeVersion_ml()
  ...

 Conflicts:
	iocore/net/UnixNet.cc
	iocore/net/UnixUDPNet.cc
	iocore/net/libinknet_stub.cc

+++ src/tscore/CompileParseRules.cc
@@ -161,7 +161,7 @@ main()
   for (c = 0; c < 256; c++) {
     fprintf(fp, "/* %3d (%c) */\t", c, (isprint(c) ? c : '?'));
     fprintf(fp, "0x%08X%c\t\t", tparseRulesCType[c], (c != 255 ? ',' : ' '));
-    fprintf(fp, "/* [%s] */\n", uint_to_binary((tparseRulesCType[c])));
+    fprintf(fp, "/* [%s] */\n", uint_to_binary((unsigned int)(tparseRulesCType[c])));
   }
   fclose(fp);
   fp = fopen("ParseRulesCTypeToUpper", "w");
